flask-app/requirements.txt
Flask
redis
prometheus_client
psycopg2-binary

